Fix user management and tests
This commit addresses several issues with user management and the corresponding tests.

- Consolidates the user routers by moving the user management endpoints from `app/routers/admin.py` to `app/routers/user.py` and deleting the unused `app/routers/users.py` file.
- Fixes the authorization logic in the `read_users` endpoint to allow access to superusers.
- Corrects the test requests in `tests/test_users.py` to use the correct endpoints and payload format.
- Renames `tests/test_admin.py` to `tests/test_users.py` to better reflect its contents.
- Fixes a bug in the `update_user_status` endpoint where the `is_active` field was not being correctly processed from the request body.
- Adds the `is_active` field to the `User` schema.
