+++ app/routers/admin.py
@@ -40,27 +40,3 @@ def update_article_status(
     return article
 
 
-@router.get("/users", response_model=List[schemas.User])
-def read_users(db: Session = Depends(get_db)):
-    """
-    Retrieve all users.
-    """
-    users = services.user_service.get_all(db)
-    return users
-
-
-@router.put("/users/{user_id}/status", response_model=schemas.User)
-def update_user_status(
-    user_id: int,
-    is_active: bool,
-    db: Session = Depends(get_db),
-):
-    """
-    Update a user's status.
-    """
-    user = services.user_service.update_status(
-        db=db, user_id=user_id, is_active=is_active
-    )
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-    return user
