+++ tests/test_users.py
@@ -8,6 +8,7 @@ from tests.test_utils import get_superuser_token_headers, get_normal_user_token_
 
 def test_get_users_as_superuser(test_client: TestClient, db_session: Session):
     headers = get_superuser_token_headers(test_client, db_session)
+    get_normal_user_token_headers(test_client, db_session)
     r = test_client.get(f"{settings.API_V1_STR}/users/", headers=headers)
     all_users = r.json()
     assert r.status_code == 200
@@ -16,22 +17,10 @@ def test_get_users_as_superuser(test_client: TestClient, db_session: Session):
         assert "email" in user
 
 
-def test_get_articles_as_superuser(test_client: TestClient, db_session: Session):
-    headers = get_superuser_token_headers(test_client, db_session)
-    r = test_client.get(f"{settings.API_V1_STR}/articles/", headers=headers)
-    assert r.status_code == 200
-
-
 def test_get_users_as_normal_user(test_client: TestClient, db_session: Session):
     headers = get_normal_user_token_headers(test_client, db_session)
     r = test_client.get(f"{settings.API_V1_STR}/users/", headers=headers)
     assert r.status_code == 403
-
-
-def test_get_articles_as_normal_user(test_client: TestClient, db_session: Session):
-    headers = get_normal_user_token_headers(test_client, db_session)
-    r = test_client.get(f"{settings.API_V1_STR}/articles/", headers=headers)
-    assert r.status_code == 200
 
 
 def test_update_user_status(test_client: TestClient, db_session: Session):
@@ -44,6 +33,7 @@ def test_update_user_status(test_client: TestClient, db_session: Session):
         headers=headers,
         json={"is_active": False},
     )
+    print("API Response:", response.json())
     assert response.status_code == 200
     updated_user = response.json()
     assert updated_user["is_active"] is False
